requirements.txt
Django~=2.0.6
google-cloud
google-cloud-vision
google-cloud-storage
django-storages
Pillow
gunicorn==19.7.1
mysqlclient
cloudstorage
